+++ tests/pass-dep/shims/libc-fs-with-isolation.stderr
@@ -1,0 +1,6 @@
+warning: `fcntl` was made to return an error due to isolation
+
+warning: `readlink` was made to return an error due to isolation
+
+warning: `$STAT` was made to return an error due to isolation
+
