+++ tests/pass-dep/shims/libc-fs.stderr
