+++ tests/pass-dep/shims/libc-fs.stdout
