+++ tests/pass-dep/shims/libc-rsfs.stdout
@@ -1,0 +1,1 @@
+hello dup fd
