+++ tests/pass/shims/fs-with-isolation.stderr
@@ -1,12 +1,8 @@
 warning: `open` was made to return an error due to isolation
 
-warning: `fcntl` was made to return an error due to isolation
-
 warning: `unlink` was made to return an error due to isolation
 
 warning: `symlink` was made to return an error due to isolation
-
-warning: `readlink` was made to return an error due to isolation
 
 warning: `$STAT` was made to return an error due to isolation
